CID: 1351700 : Result not floating point (#2434)
Description:
The result of the division is truncated to an integer (a whole number), which is usually a loss of precision in a calculation.
Type casted bricks_to_use to double.

Updates: #1060

+++ xlators/cluster/dht/src/dht-selfheal.c
@@ -1792,7 +1792,7 @@ dht_selfheal_layout_new_directory(call_frame_t *frame, loc_t *loc,
                      chunk);
     } else {
         weight_by_size = _gf_false;
-        chunk = ((unsigned long)0xffffffff) / bricks_to_use;
+        chunk = ((double)0xffffffff) / ((double)bricks_to_use);
     }
 
     start_subvol = dht_selfheal_layout_alloc_start(this, loc, layout);
